+++ common/src/main/java/net/mehvahdjukaar/sleep_tight/SleepTightClient.java
@@ -29,6 +29,7 @@ public class SleepTightClient {
     public static final ResourceLocation BED_SHEET = new ResourceLocation("textures/atlas/beds.png");
     public static final ResourceLocation ICONS = SleepTight.res("textures/gui/sleep_icons.png");
     public static final ResourceLocation BEDBUG_TEXTURE = SleepTight.res("textures/entity/bedbug.png");
+    public static final ResourceLocation SLEEPING_VILLAGER_TEXTURE = SleepTight.res("textures/entity/villager_sleeping.png");
 
     public static final Material[] HAMMOCK_TEXTURES = Arrays.stream(DyeColor.values())
             .sorted(Comparator.comparingInt(DyeColor::getId))